# Use `var` to speed up normalisation