In [this comparison](https://discourse.julialang.org/t/ann-lux-jl-explicitly-parameterized-neural-networks-in-julia/81689/5), I think that [one line](https://github.com/FluxML/Flux.jl/blob/5f17f1c32fc1cdc2c65a6f5c583f4a2734238dca/src/layers/normalise.jl#L198) accounts for Flux's extra memory use, compared to [Lux's version](https://github.com/avik-pal/Lux.jl/blob/main/src/layers/normalize.jl#L147-L148) with `var`. (Perhaps that wasn't supported earlier?). This PR fixes it:
```julia
julia> x = rand(rng, Float32, 128, 1000);

julia> @btime gradient(model -> sum(model(x)), model);
  4.906 ms (1305 allocations: 29.19 MiB)  # before
  3.136 ms (1219 allocations: 23.32 MiB)  # after

julia> v, re = destructure(model);

julia> @btime gradient(v -> sum(re(v)(x)), v);
  5.044 ms (1562 allocations: 29.47 MiB)  # before
  3.306 ms (1476 allocations: 23.60 MiB)  # after
```
compared to Lux, same machine same size:
```julia
julia> @btime gradient(p -> sum(Lux.apply(model, x, p, st)[1]), ps);
  4.069 ms (2679 allocations: 23.41 MiB)

julia> ca = ComponentArray(ps);

julia> @btime gradient(p -> sum(Lux.apply(model, x, p, st)[1]), ca);
  4.327 ms (3061 allocations: 24.62 MiB)
```