@CarloLucibello 2022-05-27
Does the gradient propagate to the keyword argument? I think that was the problem