@mcabbott 2022-05-27
> Does the gradient propagate to the keyword argument? I think that was the problem

The gradient of the keyword is correctly zero, I think:
```julia
julia> gradient(x3, m3) do x, m
         var(x; mean=m)
       end
([-0.07914716919668174, 0.04399098883554592, 0.035156180361135936], nothing)

julia> ForwardDiff.gradient(x3) do x
         var(x; mean=m3)
       end
3-element Vector{Float64}:
 -0.07914716919668174
  0.04399098883554592
  0.035156180361135936

julia> ForwardDiff.derivative(m3) do m
         var(x3; mean=m)
       end
-1.1102230246251565e-16
```
Xref https://github.com/FluxML/Zygote.jl/pull/478