@mcabbott 2022-05-27
You could save a few more copies here by making some `μ, σ2 = mean_var(x)` which computes both gradients. Or a rule for `_norm_layer_forward(x, μ, σ², ϵ)`, or better, fuse these two.