@cossio 2022-05-27
The centered second moment `<(x - m)^2>` has a minimum when `m` coincides with the mean, `m = <x>`, so the gradient is correctly zero in this case.